refactor: dedupe volme backend implementations
